+++ src/analyzer/mod.rs
@@ -17,6 +17,7 @@ pub mod dns;
 pub mod enip;
 pub mod http;
 pub mod iec104;
+pub mod iso_on_tcp;
 pub mod modbus;
 pub mod tls;